// src/app/model/common/Testata-domanda-cod-punteggio.ts
export class TestataDomandaCodPunteggio {
  nomeMinore: string;
  cognomeMinore: string;
  idDomandaIscrizione: number;
  protocolloDomandaIscrizione: string;
  statoDomandaCodice: string;
  dataConsegna: string;
  dataUltimaModifica: string;
  indirizzo: string;
  ricorrenza: string;
  codStatoValidazione: string;
}